Postgres specs using single docker container
